serial_core: Provide a transceiver interface
Before this change, to enable transceiver control for RS-485 ports,
client drivers had to implement the TIOCSRS485/TIOCGRS485 ioctls. This
change provides a common transceiver interface so that we have a common
way for all client drivers to do transceiver control.

When registering a new port, client drivers will have the option of
providing a struct txvr_ops which has callbacks for transceiver
operations. This makes it easier to have a common UART driver (like the
8250) for both RS-232 and RS-485 ports and transceiver operations can
be added for the RS-485 ports alone.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Signed-off-by: Xander Huff <xander.huff@ni.com>
Acked-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Natinst-ReviewBoard-ID: 88262
Signed-off-by: Kae Woei Kang <kae.woei.kang@emerson.com>