8250: Handle FPGA reprogramming
The 8250 driver will now subscribe to and handle fpgaperipheral
notifications. When the FPGA disappears, the port be suspended, but
not removed from the system. When the FPGA reappears, the port will be
resumed.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
(gratian: squashed with "8250: Replace 'FGPA' with 'FPGA'")
Signed-off-by: Xander Huff <xander.huff@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Natinst-ReviewBoard-ID: 156851
(cherry picked from commit 89c5032002965d3efdf2812eb2f81541698da001)
[gratian: fix conflicts with 4047b37122d1 ("serial: core: Prevent unsafe uart port access, part 1")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Signed-off-by: Kae Woei Kang <kae.woei.kang@emerson.com>